humansize.c: explicit cast to uint64_t
At this point in the code, *s is between '0' and '9' (inclusive), so (*s - '0')
is between 0 and 9 (inclusive), and can be safely cast to uint64_t.

+++ libcperciva/util/humansize.c
@@ -93,10 +93,10 @@ humansize_parse(const char * s, uint64_t * size)
 					state = -1;
 				else
 					*size *= 10;
-				if (*size > UINT64_MAX - (*s - '0'))
+				if (*size > UINT64_MAX - (uint64_t)(*s - '0'))
 					state = -1;
 				else
-					*size += *s - '0';
+					*size += (uint64_t)(*s - '0');
 				break;
 			}
